sonos-grouper/requirements.txt
soco>=0.30,<1.0
python-dotenv>=1.0,<2.0
